Fixes issues reported after upgrading SonarLint and Stylecop to their latest versions.
