+++ Aggregator.Core/Aggregator.Core.csproj
@@ -114,9 +114,16 @@
     <Compile Include="Facade\RevisionWrapper.cs" />
     <Compile Include="Interfaces\IProcessTemplateVersion.cs" />
     <Compile Include="Interfaces\IRevision.cs" />
+    <Compile Include="Interfaces\IWorkItemExposed.cs" />
+    <Compile Include="Interfaces\IWorkItemImplementation.cs" />
+    <Compile Include="Interfaces\IWorkItemRepositoryExposed.cs" />
+    <Compile Include="Monitoring\ILogger.cs" />
     <Compile Include="Monitoring\IRuleLogger.cs" />
+    <Compile Include="Monitoring\LogLevel.cs" />
     <Compile Include="Monitoring\RuleLogger.cs" />
     <Compile Include="RateLimiter.cs" />
+    <Compile Include="Script\DotNetScriptEngine{TCodeDomProvider}.cs" />
+    <Compile Include="Script\IDotNetScript.cs" />
     <Compile Include="WorkItemImplementationBase.cs" />
     <Compile Include="Configuration\Policy.cs" />
     <Compile Include="Configuration\Rule.cs" />
@@ -148,7 +155,6 @@
     <Compile Include="Context\RuntimeContext.cs" />
     <Compile Include="Script\VBNetScriptEngine.cs" />
     <Compile Include="Script\CSharpScriptEngine.cs" />
-    <Compile Include="Script\DotNetScriptEngine.cs" />
     <Compile Include="Script\ScriptEngine.cs" />
     <Compile Include="Interfaces\IFieldCollection.cs" />
     <Compile Include="Interfaces\IField.cs" />
@@ -177,10 +183,19 @@
     <Content Include="TfsAggregator.ico" />
   </ItemGroup>
   <ItemGroup>
-    <Analyzer Include="..\packages\SonarLint.1.3.0\tools\analyzers\C#\SonarLint.dll" />
-    <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta013\analyzers\dotnet\cs\StyleCop.Analyzers.dll" />
+    <Analyzer Include="..\packages\SonarLint.1.4.0\analyzers\dotnet\SonarLint.dll" />
+    <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta015\analyzers\dotnet\cs\Newtonsoft.Json.dll" />
+    <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta015\analyzers\dotnet\cs\StyleCop.Analyzers.CodeFixes.dll" />
+    <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0-beta015\analyzers\dotnet\cs\StyleCop.Analyzers.dll" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <Import Project="..\packages\StyleCop.Analyzers.1.0.0-beta015\build\StyleCop.Analyzers.targets" Condition="Exists('..\packages\StyleCop.Analyzers.1.0.0-beta015\build\StyleCop.Analyzers.targets')" />
+  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
+    <PropertyGroup>
+      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
+    </PropertyGroup>
+    <Error Condition="!Exists('..\packages\StyleCop.Analyzers.1.0.0-beta015\build\StyleCop.Analyzers.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\StyleCop.Analyzers.1.0.0-beta015\build\StyleCop.Analyzers.targets'))" />
+  </Target>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
   <Target Name="BeforeBuild">
