+++ Aggregator.Core/Interfaces/IWorkItemImplementation.cs
@@ -1,0 +1,7 @@
+﻿namespace Aggregator.Core.Interfaces
+{
+    public interface IWorkItemImplementation
+    {
+        IWorkItemLinkCollection WorkItemLinks { get; }
+    }
+}
