+++ Aggregator.Core/Interfaces/IWorkItemRepositoryExposed.cs
@@ -1,0 +1,14 @@
+namespace Aggregator.Core.Interfaces
+{
+    /// <summary>
+    /// This interface is visible to scripts.
+    /// </summary>
+    public interface IWorkItemRepositoryExposed
+    {
+        IWorkItem GetWorkItem(int workItemId);
+
+        IWorkItem MakeNewWorkItem(string projectName, string workItemTypeName);
+
+        IWorkItem MakeNewWorkItem(IWorkItem inSameProjectAs, string workItemTypeName);
+    }
+}
