+++ Aggregator.Core/Monitoring/ILogEvents.cs
@@ -11,26 +11,6 @@ using Microsoft.TeamFoundation.WorkItemTracking.Client;
 namespace Aggregator.Core.Monitoring
 {
     /// <summary>
-    /// Levels of logging.
-    /// </summary>
-    /// <remarks>While this enumeration is not used within Core, it is read by the configuration class <see cref="Aggregator.Core.Configuration.TFSAggregatorSettings"/>.</remarks>
-    public enum LogLevel
-    {
-        Critical = 1,
-        Error = 2,
-        Warning = 3,
-        Information = 5,
-        Normal = Information,
-        Verbose = 10,
-        Diagnostic = 99,
-    }
-
-    public interface ILogger
-    {
-        LogLevel MinimumLogLevel { get; set; }
-    }
-
-    /// <summary>
     /// Core Clients will be called on this interface to log events and errors.
     /// </summary>
     /// <remarks>The methods must *not* raise exceptions.</remarks>
