+++ Aggregator.Core/Monitoring/ILogger.cs
@@ -1,0 +1,7 @@
+﻿namespace Aggregator.Core.Monitoring
+{
+    public interface ILogger
+    {
+        LogLevel MinimumLogLevel { get; set; }
+    }
+}
