+++ Aggregator.Core/Script/IDotNetScript.cs
@@ -1,0 +1,10 @@
+using Aggregator.Core.Interfaces;
+using Aggregator.Core.Monitoring;
+
+namespace Aggregator.Core.Script
+{
+    public interface IDotNetScript
+    {
+        object RunScript(IWorkItemExposed self, IWorkItemRepositoryExposed store, IRuleLogger scriptLogger);
+    }
+}
