+++ Aggregator.Core/packages.config
@@ -9,8 +9,8 @@
   <package id="Microsoft.VisualStudio.Services.InteractiveClient" version="14.83.0" targetFramework="net45" />
   <package id="Microsoft.WindowsAzure.ConfigurationManager" version="1.7.0.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="6.0.5" targetFramework="net45" />
-  <package id="SonarLint" version="1.3.0" targetFramework="net45" developmentDependency="true" />
-  <package id="StyleCop.Analyzers" version="1.0.0-beta013" targetFramework="net45" developmentDependency="true" />
+  <package id="SonarLint" version="1.4.0" targetFramework="net45" developmentDependency="true" />
+  <package id="StyleCop.Analyzers" version="1.0.0-beta015" targetFramework="net45" developmentDependency="true" />
   <package id="System.IdentityModel.Tokens.Jwt" version="4.0.0" targetFramework="net45" />
   <package id="System.Management.Automation_PowerShell_3.0" version="6.3.9600.17400" targetFramework="net45" />
   <package id="WindowsAzure.ServiceBus" version="2.5.1.0" targetFramework="net45" />
