+++ ManualTests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="StyleCop.Analyzers" version="1.0.0-beta013" targetFramework="net452" developmentDependency="true" />
+  <package id="SonarLint" version="1.4.0" targetFramework="net452" developmentDependency="true" />
+  <package id="StyleCop.Analyzers" version="1.0.0-beta015" targetFramework="net452" developmentDependency="true" />
 </packages>
