+++ TFS-Aggregator-2.ruleset
@@ -66,6 +66,7 @@
   </Rules>
   <Rules AnalyzerId="SonarLint" RuleNamespace="SonarLint">
     <Rule Id="S2228" Action="None" />
+    <Rule Id="S2372" Action="Info" />
     <Rule Id="S3215" Action="None" />
   </Rules>
   <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
@@ -79,7 +80,6 @@
     <Rule Id="SA1600" Action="None" />
     <Rule Id="SA1601" Action="None" />
     <Rule Id="SA1602" Action="None" />
-    <Rule Id="SA1609" Action="None" />
     <Rule Id="SA1610" Action="None" />
     <Rule Id="SA1611" Action="None" />
     <Rule Id="SA1615" Action="None" />
